Time-Stamp Changes
No conversion should be needed here. This makes the time-stamps inaccurate. If a user wants JST time it's more convenient to convert it using the already provided universal time.
(Is two seconds being subtracted because of the host? I'm confused as sometimes EQ's are posted a second after the minute mark..)